+++ app/concepts/stellar_base/deposit_requests/operations/create.rb
@@ -54,8 +54,7 @@ module StellarBase
           options["model"].min_amount = 0.0
 
           # Make Deposits free unless we want it configured
-          #
-          # TODO: this should come from `CallFeeFrom.(details[:fee_from)`
+          # TODO: this should come from `CallDepositFeeFrom.(params, details)`
           options["model"].fee_fixed = details[:fee_fixed] || 0.0
           options["model"].fee_percent = details[:fee_percent] || 0.0
         end