+++ app/services/stellar_base/fee_requests/call_fee_from.rb
@@ -2,25 +2,25 @@ module StellarBase
   module FeeRequests
     class CallFeeFrom
 
-      FEE_FROM_ERROR = "`fee_from` must be a string of an object " \
+      FIXED_FEE_QUOTE_FROM_ERROR = "`fixed_fee_quote_from` must be a string of an object " \
         "that responds to `.call` or the object itself".freeze
 
-      def self.call(fee_request:, fee_from:)
-        return 0 if fee_from.nil?
-        callback = callback_from(fee_from)
+      def self.call(fee_request:, fixed_fee_quote_from:)
+        return 0 if fixed_fee_quote_from.nil?
+        callback = callback_from(fixed_fee_quote_from)
         callback.(fee_request)
       end
 
-      def self.callback_from(fee_from)
-        if fee_from.respond_to?(:constantize)
-          fee_from = fee_from.constantize
+      def self.callback_from(fixed_fee_quote_from)
+        if fixed_fee_quote_from.respond_to?(:constantize)
+          fixed_fee_quote_from = fixed_fee_quote_from.constantize
         end
 
-        if !fee_from.respond_to?(:call)
-          fail ArgumentError, FEE_FROM_ERROR
+        if !fixed_fee_quote_from.respond_to?(:call)
+          fail ArgumentError, FIXED_FEE_QUOTE_FROM_ERROR
         end
 
-        fee_from
+        fixed_fee_quote_from
       end
 
     end