+++ spec/dummy/app/services/get_withdraw_fixed_fee_quote_from.rb
@@ -1,4 +1,4 @@
-class GetWithdrawFeeFrom
+class GetWithdrawFixedFeeQuoteFrom
 
   def self.call(fee_request)
     0.0001