+++ spec/services/stellar_base/find_asset_details_spec.rb
@@ -46,7 +46,7 @@ module StellarBase
           issuer: ENV["ISSUER_ADDRESS"],
           fee_fixed: 0.01,
           max_amount_from: GetMaxAmount.to_s,
-          fee_from: GetWithdrawFeeFrom,
+          fixed_fee_quote_from: GetWithdrawFixedFeeQuoteFrom,
         })
       end
     end