+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Rename `fee_fixed_from` to `fee_fixed_quote_from` to reflect that this is used in the `/fee` endpoint for quoting
+- Remove `fee_network`: this is not a Stellar standard. Any apps that use this gem should determine and/or store the network fee in its own records.
 
 ### Added
 - Developer sets `fee_fixed_from` in withdrawable asset config to customize how fees on withdrawals are set